requirements.txt
arabic-reshaper==2.1.0
backcall==0.1.0
backtrader==1.9.74.123
backtrader-plotting==1.1.0
beautifulsoup4==4.9.0
bokeh==2.0.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
cycler==0.10.0
decorator==4.4.2
dropbox==9.4.0
empyrical==0.5.3
future==0.18.2
graphviz==0.13.2
html5lib==1.0.1
idna==2.8
ipython==7.13.0
ipython-genutils==0.2.0
jdatetime==3.6.2
jedi==0.16.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.1.0
lxml==4.5.2
markdown2==2.3.8
MarkupSafe==1.1.1
matplotlib==3.0.3
mechanize==0.4.5
MetaTrader5==5.0.33
mplfinance==0.12.7a0
multitasking==0.0.9
numpy==1.18.1
objgraph==3.4.1
packaging==20.3
pandas==0.25.3
pandas-datareader==0.8.1
parso==0.6.2
patsy==0.5.1
persiantools==1.4.1
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.12.0
prompt-toolkit==3.0.4
pyfolio==0.9.2
Pygments==2.6.1
pyparsing==2.4.6
python-bidi==0.4.2
python-dateutil==2.8.1
pytse==1.2.1
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
retrying==1.3.3
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.9.1
selenium==3.141.0
six==1.14.0
soupsieve==2.0
statsmodels==0.12.0
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.8
wcwidth==0.1.9
webdriver-manager==3.2.1
webencodings==0.5.1
yfinance==0.1.55
